pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "maria eduarda", email = "mariaeduarda.gerotti@gmail.com"},
]
dependencies = [
    "django>=4.2.1",
    "djangorestframework>=3.14.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
